split componets in files
