+++ play-nine-app/src/answer/Answer.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './Answer.css';
+import 'bootstrap/dist/css/bootstrap.css';
+
+const Answer = (props) => {
+    return(
+        <div className="col-5">
+            <span>5</span>
+            <span>6</span>
+        </div>
+    );
+}
+
+export default Answer;
