+++ play-nine-app/src/game/Game.js
@@ -1,55 +1,11 @@
 import React from 'react';
 import './Game.css';
 import 'bootstrap/dist/css/bootstrap.css';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import _ from 'lodash';
-
-const Stars = (props) => {
-    const numberOfStars = 1 + Math.floor(Math.random()*9);
-
-    // let stars = [];
-    // for(let i=0;i<numberOfStars;i++){
-    //     stars.push(<FontAwesomeIcon key={i} icon="star" className="star" />)
-    // }
-    return(
-        <div className="col-5">
-           {_.range(numberOfStars).map(i => 
-                <FontAwesomeIcon key={i} icon="star" className="star" />
-            )}
-        </div>
-    );
-}
-
-const Button = (props) => {
-    return(
-        <div className="col-2">
-            <button type="button" class="btn btn-warning">=</button>
-        </div>
-    );
-}
-
-const Answer = (props) => {
-    return(
-        <div className="col-5">
-            <span>5</span>
-            <span>6</span>
-        </div>
-    );
-}
-
-const Numbers = (props) => {
-    return(
-        <div className="card text-center">
-            <div>
-                {Numbers.list.map((number, i) => 
-                    <span key={i}>{number}</span>
-                )}
-            </div>
-        </div>
-    );
-}
-
-Numbers.list = _.range(1,10);
+import Stars from '../starts/Stars';
+import Button from '../button/Button';
+import Answer from '../answer/Answer';
+import Numbers from '../numbers/Numbers';
 
 class Game extends React.Component{
     render() {
