+++ play-nine-app/src/numbers/Numbers.css
@@ -1,0 +1,11 @@
+.selected {
+    background-color: #eee;
+    color: #ddd;
+    cursor: not-allowed;
+}
+
+.used {
+    background-color: #aaddaa;
+    color: #99bb99;
+    cursor: not-allowed;
+}
