+++ play-nine-app/src/numbers/Numbers.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import './Numbers.css';
+import 'bootstrap/dist/css/bootstrap.css';
+import _ from 'lodash';
+
+const Numbers = (props) => {
+    return(
+        <div className="card text-center">
+            <div>
+                {Numbers.list.map((number, i) => 
+                    <span key={i}>{number}</span>
+                )}
+            </div>
+        </div>
+    );
+}
+
+Numbers.list = _.range(1,10);
+
+export default Numbers;
