+++ play-nine-app/src/starts/Stars.css
@@ -1,0 +1,4 @@
+.star {
+    margin: 0.5em;
+    font-size: 24px;
+}
